more javadocs
